--- a/deno/lib/__tests__/promise.test.ts
+++ b/deno/lib/__tests__/promise.test.ts
@@ -4,6 +4,7 @@ const test = Deno.test;
 
 import { util } from "../helpers/util.ts";
 import * as z from "../index.ts";
+import { ZodError } from "../ZodError.ts";
 
 const promSchema = z.promise(
   z.object({
@@ -37,7 +38,7 @@ test("promise parsing success 2", () => {
 test("promise parsing fail", async () => {
   const bad = promSchema.parse(Promise.resolve({ name: "Bobby", age: "10" }));
   // return await expect(bad).resolves.toBe({ name: 'Bobby', age: '10' });
-  return await expect(bad).rejects.toBeInstanceOf(Error);
+  return await expect(bad).rejects.toBeInstanceOf(ZodError);
   // done();
 });
 
@@ -45,7 +46,7 @@ test("promise parsing fail 2", async () => {
   const failPromise = promSchema.parse(
     Promise.resolve({ name: "Bobby", age: "10" })
   );
-  await expect(failPromise).rejects.toBeInstanceOf(Error);
+  await expect(failPromise).rejects.toBeInstanceOf(ZodError);
   // done();/z
 });
 
@@ -74,7 +75,7 @@ test("async function fail", async () => {
   const validatedFunction = asyncFunction.implement(() => {
     return Promise.resolve("asdf" as any);
   });
-  await expect(validatedFunction()).rejects.toBeInstanceOf(Error);
+  await expect(validatedFunction()).rejects.toBeInstanceOf(ZodError);
 });
 
 test("async promise parsing", () => {