--- a/package.json
+++ b/package.json
@@ -78,5 +78,8 @@
       "yarn fix:lint",
       "yarn fix:format"
     ]
+  },
+  "dependencies": {
+    "nanobench": "^2.1.1"
   }
 }