--- a/src/__tests__/promise.test.ts
+++ b/src/__tests__/promise.test.ts
@@ -3,6 +3,7 @@ import { expect, test } from "@jest/globals";
 
 import { util } from "../helpers/util";
 import * as z from "../index";
+import { ZodError } from "../ZodError";
 
 const promSchema = z.promise(
   z.object({
@@ -36,7 +37,7 @@ test("promise parsing success 2", () => {
 test("promise parsing fail", async () => {
   const bad = promSchema.parse(Promise.resolve({ name: "Bobby", age: "10" }));
   // return await expect(bad).resolves.toBe({ name: 'Bobby', age: '10' });
-  return await expect(bad).rejects.toBeInstanceOf(Error);
+  return await expect(bad).rejects.toBeInstanceOf(ZodError);
   // done();
 });
 
@@ -44,7 +45,7 @@ test("promise parsing fail 2", async () => {
   const failPromise = promSchema.parse(
     Promise.resolve({ name: "Bobby", age: "10" })
   );
-  await expect(failPromise).rejects.toBeInstanceOf(Error);
+  await expect(failPromise).rejects.toBeInstanceOf(ZodError);
   // done();/z
 });
 
@@ -73,7 +74,7 @@ test("async function fail", async () => {
   const validatedFunction = asyncFunction.implement(() => {
     return Promise.resolve("asdf" as any);
   });
-  await expect(validatedFunction()).rejects.toBeInstanceOf(Error);
+  await expect(validatedFunction()).rejects.toBeInstanceOf(ZodError);
 });
 
 test("async promise parsing", () => {