… coverage.svg
… yarn.lock