See colinhacks/zod#292.

Benchmarks for `z.string`:

colinhacks/zod@v3
```
NANOBENCH version 2
> node_modules/.bin/ts-node src/benchmarks/parser.ts

# z.string(): empty string
ok ~1.64 s (1 s + 640254015 ns)

# z.string(): null
ok ~2.23 s (2 s + 233349960 ns)

all benchmarks completed
ok ~3.87 s (3 s + 873603975 ns)
```

zod-contrib/zod
```
NANOBENCH version 2
> node_modules/.bin/ts-node src/benchmarks/parser.ts

# z.string(): empty string
ok ~244 ms (0 s + 244489747 ns)

# z.string(): null
ok ~1.06 s (1 s + 59004720 ns)

all benchmarks completed
ok ~1.3 s (1 s + 303494467 ns)
```